requirements.txt
Django==1.4.1
MySQL-python==1.2.3c1
South==0.7.6
django-pagination==1.0.7
-e git://github.com/alex/django-taggit.git@36f6dabcf10e27c7d9442a94243d4189f2a4f121#egg=django_taggit-dev
django-tinymce==1.5.1b2
wsgiref==0.1.2
